Separated Data and Presentation - Export to CSV
Separated Data from Presentation in hopes of speeding things up, but the fact we are asking for data is slowing things down.  Might be able to batch request together to speed it up... Added the ability to export the run data into a CSV file for excel or sheets use.  Should be pretty straight forward to add more data to the output if needed.  I'll very likely do one more update with more data.